# openapi.json
{
  "components": {
    "schemas": {
...
      "BinaryDataStreamingDataHandlerParameters": {
        "additionalProperties": false,
        "description": "Configuration parameters for the Binary Data Streaming Data Handler\n\nThis data handler forwards serialized byte objects to one or more remote\nendpoints over a ZMQ PUSH socket\n\nAttributes:\n\n    type: Discriminator field, must be ``\"BinaryDataStreamingDataHandler\"``\n\n    urls: List of endpoint URLs to bind to (server mode) or connect to\n        (client mode)\n\n    distribute: Boolean, if True: round robin connect the ranks, False: all ranks connect to all urls and send the same data\n\n    buffer: buffer size, if set to 0 the OS default is used\n\n    role: Whether this node acts as the ZMQ ``\"server\"`` (binds) or\n        ``\"client\"`` (connects). Defaults to ``\"server\"``\n\n    library: Underlying transport library to use. Currently only ``\"zmq\"``\n        is supported. Defaults to ``\"zmq\"``\n\n    socket_type: Socket pattern to use. Currently only ``\"push\"`` is\n        supported. Defaults to ``\"push\"``",
        "properties": {
          "buffer": {
            "default": 0,
            "title": "Buffer",
            "type": "integer"
          },
          "distribute": {
            "default": true,
            "title": "Distribute",
            "type": "boolean"
          },
          "library": {
            "const": "zmq",
            "default": "zmq",
            "title": "Library",
            "type": "string"
          },
          "role": {
            "default": "client",
            "enum": [
              "server",
              "client"
            ],
            "title": "Role",
            "type": "string"
          },
          "socket_type": {
            "const": "push",
            "default": "push",
            "title": "Socket Type",
            "type": "string"
          },
          "type": {
            "const": "BinaryDataStreamingDataHandler",
            "title": "Type",
            "type": "string"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "title": "Urls",
            "type": "array"
          }
        },
        "required": [
          "type",
          "urls"
        ],
        "title": "BinaryDataStreamingDataHandlerParameters",
        "type": "object"
      },
      "BinaryFileWritingDataHandlerParameters": {
        "additionalProperties": false,
        "description": "Configuration parameters for the Binary File Writing Data Handler\n\nThis data handler writes each serialized byte object to a separate file on\nthe filesystem\n\nAttributes:\n\n    type: Discriminator field, must be ``\"BinaryFileWritingDataHandler\"``\n\n    file_prefix: Optional string prepended to every output filename,\n        separated from the rest of the name by an underscore. Defaults to\n        ``\"\"`` (no prefix)\n\n    file_suffix: File extension used for output files, without the leading\n        dot. Defaults to ``\"h5\"``\n\n    write_directory: Directory in which output files are created. The\n        directory is created (including parents) if it does not already\n        exist. Defaults to the current working directory",
        "properties": {
          "file_prefix": {
            "default": "",
            "title": "File Prefix",
            "type": "string"
          },
          "file_suffix": {
            "default": "h5",
            "title": "File Suffix",
            "type": "string"
          },
          "type": {
            "const": "BinaryFileWritingDataHandler",
            "title": "Type",
            "type": "string"
          },
          "write_directory": {
            "format": "path",
            "title": "Write Directory",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "BinaryFileWritingDataHandlerParameters",
        "type": "object"
      },
      "CacheMode": {
        "enum": [
          "per_transfer",
          "shared"
        ],
        "title": "CacheMode",
        "type": "string"
      },
      "ConstValueParameters": {
        "additionalProperties": false,
        "description": "Parameters for ConstValue class",
        "properties": {
          "dtype": {
            "title": "Dtype",
            "type": "string"
          },
          "type": {
            "const": "ConstValue",
            "title": "Type",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "type",
          "value",
          "dtype"
        ],
        "title": "ConstValueParameters",
        "type": "object"
      },
      "ConsumerConnectionInfo": {
        "description": "The cache push socket the consumer connects to.",
        "properties": {
          "host": {
            "title": "Host",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "uri": {
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "host",
          "port",
          "uri"
        ],
        "title": "ConsumerConnectionInfo",
        "type": "object"
      },
...
      "HDF5BinarySerializerParameters": {
        "additionalProperties": false,
        "description": "Configuration parameters for the HDF5 binary serializer\n\nThis serializer encodes a batch of event data arrays into an in-memory\nHDF5 file. Optional compression can be applied to each dataset\n\nAttributes:\n\n    type: Discriminator field, must be ``\"HDF5BinarySerializer\"``\n\n    compression_level: Compression level passed to the chosen algorithm.\n        Interpretation depends on the algorithm. Defaults to ``3``\n\n    compression: Compression algorithm to use. Supported values are\n        ``\"gzip\"``, ``\"gzip_with_shuffle\"``, ``\"bitshuffle_with_lz4\"``,\n        ``\"bitshuffle_with_zstd\"``, and ``\"zfp\"``. Set to ``None`` to\n        disable compression. Defaults to ``None``\n\n    fields: Dictionary storing the mapping from data source name to the\n        HDF5 dataset path under which that source's data will be stored",
        "properties": {
          "compression": {
            "anyOf": [
              {
                "enum": [
                  "gzip",
                  "gzip_with_shuffle",
                  "bitshuffle_with_lz4",
                  "bitshuffle_with_zstd",
                  "zfp"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compression"
          },
          "compression_level": {
            "default": 0,
            "title": "Compression Level",
            "type": "integer"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "type": {
            "const": "HDF5BinarySerializer",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "fields"
        ],
        "title": "HDF5BinarySerializerParameters",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogReadMode": {
        "enum": [
          "head",
          "tail"
        ],
        "title": "LogReadMode",
        "type": "string"
      },
      "LogStream": {
        "enum": [
          "cache",
          "producer_stdout",
          "producer_stderr"
        ],
        "title": "LogStream",
        "type": "string"
      },
      "Message": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "Message",
        "type": "object"
      },
      "Parameters": {
        "additionalProperties": false,
        "description": "Top-level configuration parameters for an lclstreamer run\n\nThis model aggregates all sub-component configuration into a single object\n\nAttributes:\n\n    source_identifier: A string that uniquely identifies the data source\n        (e.g. a psana data source string)\n\n    skip_incomplete_events: When ``True``, events for which one or more\n        data sources returned no data are silently dropped from the stream\n\n    event_source: Configuration for the event source\n\n    data_sources: Mapping from arbitrary data source names to the\n        configuration parameters of each data source\n\n    processing_pipeline: Configuration for the Processing Pipeline applied\n        to the event stream\n\n    data_serializer: Configuration for the serializer that converts\n        processed events into byte objects\n\n    data_handlers: Ordered list of data handler configurations; each\n        handler receives the serialized byte object in turn",
        "properties": {
          "data_handlers": {
            "items": {
              "discriminator": {
                "mapping": {
                  "BinaryDataStreamingDataHandler": "#/components/schemas/BinaryDataStreamingDataHandlerParameters",
                  "BinaryFileWritingDataHandler": "#/components/schemas/BinaryFileWritingDataHandlerParameters"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BinaryDataStreamingDataHandlerParameters"
                },
                {
                  "$ref": "#/components/schemas/BinaryFileWritingDataHandlerParameters"
                }
              ]
            },
            "title": "Data Handlers",
            "type": "array"
          },
          "data_serializer": {
            "discriminator": {
              "mapping": {
                "HDF5BinarySerializer": "#/components/schemas/HDF5BinarySerializerParameters",
                "SimplonBinarySerializer": "#/components/schemas/SimplonBinarySerializerParameters"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/HDF5BinarySerializerParameters"
              },
              {
                "$ref": "#/components/schemas/SimplonBinarySerializerParameters"
              }
            ],
            "title": "Data Serializer"
          },
          "data_sources": {
            "additionalProperties": {
              "discriminator": {
                "mapping": {
                  "ConstValue": "#/components/schemas/ConstValueParameters",
                  "GenericRandomNumpyArray": "#/components/schemas/GenericRandomNumpyArrayParameters",
                  "Psana1DetectorInterface": "#/components/schemas/Psana1DetectorInterfaceParameters",
                  "Psana1Timestamp": "#/components/schemas/Psana1TimestampParameters",
                  "Psana2DetectorInterface": "#/components/schemas/Psana2DetectorInterfaceParameters",
                  "Psana2RunInfo": "#/components/schemas/Psana2RunInfoParameters",
                  "Psana2Timestamp": "#/components/schemas/Psana2TimestampParameters",
                  "SourceIdentifier": "#/components/schemas/SourceIdentifierParameters"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/GenericRandomNumpyArrayParameters"
                },
                {
                  "$ref": "#/components/schemas/ConstValueParameters"
                },
                {
                  "$ref": "#/components/schemas/Psana1DetectorInterfaceParameters"
                },
                {
                  "$ref": "#/components/schemas/Psana2DetectorInterfaceParameters"
                },
                {
                  "$ref": "#/components/schemas/Psana1TimestampParameters"
                },
                {
                  "$ref": "#/components/schemas/Psana2TimestampParameters"
                },
                {
                  "$ref": "#/components/schemas/SourceIdentifierParameters"
                },
                {
                  "$ref": "#/components/schemas/Psana2RunInfoParameters"
                }
              ]
            },
            "title": "Data Sources",
            "type": "object"
          },
          "event_source": {
            "discriminator": {
              "mapping": {
                "InternalEventSource": "#/components/schemas/InternalEventSourceParameters",
                "Psana1EventSource": "#/components/schemas/Psana1EventSourceParameters",
                "Psana2EventSource": "#/components/schemas/Psana2EventSourceParameters"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/InternalEventSourceParameters"
              },
              {
                "$ref": "#/components/schemas/Psana1EventSourceParameters"
              },
              {
                "$ref": "#/components/schemas/Psana2EventSourceParameters"
              }
            ],
            "title": "Event Source"
          },
          "processing_pipeline": {
            "discriminator": {
              "mapping": {
                "BatchProcessingPipeline": "#/components/schemas/BatchProcessingPipelineParameters",
                "PeaknetPreprocessingPipeline": "#/components/schemas/PeaknetPreprocessingPipelineParameters"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BatchProcessingPipelineParameters"
              },
              {
                "$ref": "#/components/schemas/PeaknetPreprocessingPipelineParameters"
              }
            ],
            "title": "Processing Pipeline"
          },
          "skip_incomplete_events": {
            "title": "Skip Incomplete Events",
            "type": "boolean"
          },
          "source_identifier": {
            "title": "Source Identifier",
            "type": "string"
          }
        },
        "required": [
          "source_identifier",
          "skip_incomplete_events",
          "event_source",
          "data_sources",
          "processing_pipeline",
          "data_serializer",
          "data_handlers"
        ],
        "title": "Parameters",
        "type": "object"
      },
...
      "Psana1TimestampParameters": {
        "additionalProperties": false,
        "description": "Parameters for psana1 timestamp interface",
        "properties": {
          "type": {
            "const": "Psana1Timestamp",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "Psana1TimestampParameters",
        "type": "object"
      },
...
      "Psana2RunInfoParameters": {
        "additionalProperties": false,
        "description": "Parameters for run info data source interface",
        "properties": {
          "type": {
            "const": "Psana2RunInfo",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "Psana2RunInfoParameters",
        "type": "object"
      },
...
      "TransferCreate": {
        "properties": {
          "cache_mode": {
            "$ref": "#/components/schemas/CacheMode",
            "default": "per_transfer"
          },
          "parameters": {
            "$ref": "#/components/schemas/Parameters"
          }
        },
        "required": [
          "parameters"
        ],
        "title": "TransferCreate",
        "type": "object"
      },
      "TransferDetail": {
        "properties": {
          "cache_mode": {
            "$ref": "#/components/schemas/CacheMode"
          },
          "connection_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsumerConnectionInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "experiment": {
            "title": "Experiment",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "run": {
            "title": "Run",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TransferState"
          },
          "transitions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TransitionPublic"
            },
            "title": "Transitions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "created_at",
          "requested_by",
          "state",
          "experiment",
          "run",
          "cache_mode"
        ],
        "title": "TransferDetail",
        "type": "object"
      },
      "TransferLogIndex": {
        "properties": {
          "streams": {
            "items": {
              "$ref": "#/components/schemas/TransferLogStreamInfo"
            },
            "title": "Streams",
            "type": "array"
          },
          "transfer_id": {
            "format": "uuid",
            "title": "Transfer Id",
            "type": "string"
          }
        },
        "required": [
          "transfer_id",
          "streams"
        ],
        "title": "TransferLogIndex",
        "type": "object"
      },
      "TransferLogStreamInfo": {
        "description": "One log stream's resolved location and (best-effort) availability.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "modified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "path": {
            "format": "path",
            "title": "Path",
            "type": "string"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "stream": {
            "$ref": "#/components/schemas/LogStream"
          }
        },
        "required": [
          "stream",
          "path",
          "available"
        ],
        "title": "TransferLogStreamInfo",
        "type": "object"
      },
      "TransferPublic": {
        "properties": {
          "cache_mode": {
            "$ref": "#/components/schemas/CacheMode"
          },
          "connection_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsumerConnectionInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "experiment": {
            "title": "Experiment",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "run": {
            "title": "Run",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TransferState"
          }
        },
        "required": [
          "id",
          "created_at",
          "requested_by",
          "state",
          "experiment",
          "run",
          "cache_mode"
        ],
        "title": "TransferPublic",
        "type": "object"
      },
      "TransferState": {
        "enum": [
          "provisioning",
          "ready",
          "canceling",
          "canceled",
          "completed",
          "failed"
        ],
        "title": "TransferState",
        "type": "string"
      },
      "TransfersPublic": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TransferPublic"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "count"
        ],
        "title": "TransfersPublic",
        "type": "object"
      },
      "TransitionPublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransitionSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/TransferState"
          }
        },
        "required": [
          "created_at",
          "state"
        ],
        "title": "TransitionPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "JWKSAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Durable lclstreamer-based data transfers.",
    "title": "LCLStream API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/transfers/": {
      "get": {
        "operationId": "get_transfers_transfers__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TransferState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransfersPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWKSAuth": []
          }
        ],
        "summary": "Get Transfers",
        "tags": [
          "transfers"
        ]
      },
      "post": {
        "operationId": "new_transfer_transfers__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWKSAuth": []
          }
        ],
        "summary": "New Transfer",
        "tags": [
          "transfers"
        ]
      }
    },
    "/transfers/{transfer_id}": {
      "delete": {
        "operationId": "cancel_transfer_transfers__transfer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "transfer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transfer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWKSAuth": []
          }
        ],
        "summary": "Cancel Transfer",
        "tags": [
          "transfers"
        ]
      },
      "get": {
        "operationId": "get_transfer_transfers__transfer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transfer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transfer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWKSAuth": []
          }
        ],
        "summary": "Get Transfer",
        "tags": [
          "transfers"
        ]
      }
    },
    "/transfers/{transfer_id}/logs": {
      "get": {
        "operationId": "get_transfer_logs_transfers__transfer_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "transfer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transfer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferLogIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWKSAuth": []
          }
        ],
        "summary": "Get Transfer Logs",
        "tags": [
          "transfers"
        ]
      }
    },
    "/transfers/{transfer_id}/logs/{stream}": {
      "get": {
        "operationId": "get_transfer_log_transfers__transfer_id__logs__stream__get",
        "parameters": [
          {
            "in": "path",
            "name": "transfer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transfer Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stream",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LogStream"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LogReadMode",
              "default": "tail"
            }
          },
          {
            "in": "query",
            "name": "lines",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Lines",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "bytes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bytes"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "JWKSAuth": []
          }
        ],
        "summary": "Get Transfer Log",
        "tags": [
          "transfers"
        ]
      }
    }
  }
}